refactor(plugin): decouple plugin framework from broker internals (#558)
Move Plugin trait, CommandInfo, ParamSchema, ParamType, PluginResult,
CommandContext, RoomMetadata, UserInfo, and BoxFuture to room-protocol.
Add MessageWriter and HistoryAccess async traits as abstractions over
broker broadcast/persistence and history access.

Concrete ChatWriter and HistoryReader implementations moved from
plugin/mod.rs to a new plugin/bridge.rs module — the only plugin
submodule that imports from crate::broker. plugin/mod.rs now has
zero broker imports.

CommandContext uses Box<dyn MessageWriter> and Box<dyn HistoryAccess>
instead of concrete types, enabling external crates to implement
Plugin by depending on room-protocol alone.

Closes #558
