+++ crates/room-cli/src/plugin/mod.rs
@@ -1,340 +1,21 @@
+pub mod bridge;
 pub mod queue;
 pub mod stats;
 pub mod taskboard;
 
-use std::{
-    collections::HashMap,
-    future::Future,
-    path::{Path, PathBuf},
-    pin::Pin,
-    sync::{
-        atomic::{AtomicU64, Ordering},
-        Arc,
-    },
+use std::{collections::HashMap, path::Path};
+
+// Re-export all plugin framework types from room-protocol so that existing
+// imports from `crate::plugin::*` continue to work without changes.
+pub use room_protocol::plugin::{
+    BoxFuture, CommandContext, CommandInfo, HistoryAccess, MessageWriter, ParamSchema, ParamType,
+    Plugin, PluginResult, RoomMetadata, UserInfo,
 };
 
-use chrono::{DateTime, Utc};
-
-use crate::{
-    broker::{
-        fanout::broadcast_and_persist,
-        state::{ClientMap, StatusMap},
-    },
-    history,
-    message::{make_event, make_system, EventType, Message},
-};
-
-/// Boxed future type used by [`Plugin::handle`] for dyn compatibility.
-pub type BoxFuture<'a, T> = Pin<Box<dyn Future<Output = T> + Send + 'a>>;
-
-// ── Plugin trait ────────────────────────────────────────────────────────────
-
-/// A plugin that handles one or more `/` commands and/or reacts to room
-/// lifecycle events.
-///
-/// Implement this trait and register it with [`PluginRegistry`] to add
-/// custom commands to a room broker. The broker dispatches matching
-/// `Message::Command` messages to the plugin's [`handle`](Plugin::handle)
-/// method, and calls [`on_user_join`](Plugin::on_user_join) /
-/// [`on_user_leave`](Plugin::on_user_leave) when users enter or leave.
-///
-/// Only [`name`](Plugin::name) and [`handle`](Plugin::handle) are required.
-/// All other methods have no-op / empty-vec defaults so that adding new
-/// lifecycle hooks in future releases does not break existing plugins.
-pub trait Plugin: Send + Sync {
-    /// Unique identifier for this plugin (e.g. `"stats"`, `"help"`).
-    fn name(&self) -> &str;
-
-    /// Commands this plugin handles. Each entry drives `/help` output
-    /// and TUI autocomplete.
-    ///
-    /// Defaults to an empty vec for plugins that only use lifecycle hooks
-    /// and do not register any commands.
-    fn commands(&self) -> Vec<CommandInfo> {
-        vec![]
-    }
-
-    /// Handle an invocation of one of this plugin's commands.
-    ///
-    /// Returns a boxed future for dyn compatibility (required because the
-    /// registry stores `Box<dyn Plugin>`).
-    fn handle(&self, ctx: CommandContext) -> BoxFuture<'_, anyhow::Result<PluginResult>>;
-
-    /// Called after a user joins the room. The default is a no-op.
-    ///
-    /// Invoked synchronously during the join broadcast path. Implementations
-    /// must not block — spawn a task if async work is needed.
-    fn on_user_join(&self, _user: &str) {}
-
-    /// Called after a user leaves the room. The default is a no-op.
-    ///
-    /// Invoked synchronously during the leave broadcast path. Implementations
-    /// must not block — spawn a task if async work is needed.
-    fn on_user_leave(&self, _user: &str) {}
-}
-
-// ── CommandInfo ─────────────────────────────────────────────────────────────
-
-/// Describes a single command for `/help` and autocomplete.
-#[derive(Debug, Clone)]
-pub struct CommandInfo {
-    /// Command name without the leading `/`.
-    pub name: String,
-    /// One-line description shown in `/help` and autocomplete.
-    pub description: String,
-    /// Usage string (e.g. `"/stats [last N]"`).
-    pub usage: String,
-    /// Typed parameter schemas for validation and autocomplete.
-    pub params: Vec<ParamSchema>,
-}
-
-// ── Typed parameter schema ─────────────────────────────────────────────────
-
-/// Schema for a single command parameter — drives validation, `/help` output,
-/// and TUI argument autocomplete.
-#[derive(Debug, Clone)]
-pub struct ParamSchema {
-    /// Display name (e.g. `"username"`, `"count"`).
-    pub name: String,
-    /// What kind of value this parameter accepts.
-    pub param_type: ParamType,
-    /// Whether the parameter must be provided.
-    pub required: bool,
-    /// One-line description shown in `/help <command>`.
-    pub description: String,
-}
-
-/// The kind of value a parameter accepts.
-#[derive(Debug, Clone, PartialEq)]
-pub enum ParamType {
-    /// Free-form text (no validation beyond presence).
-    Text,
-    /// One of a fixed set of allowed values.
-    Choice(Vec<String>),
-    /// An online username — TUI shows the mention picker.
-    Username,
-    /// An integer, optionally bounded.
-    Number { min: Option<i64>, max: Option<i64> },
-}
-
-// ── CommandContext ───────────────────────────────────────────────────────────
-
-/// Context passed to a plugin's `handle` method.
-pub struct CommandContext {
-    /// The command name that was invoked (without `/`).
-    pub command: String,
-    /// Arguments passed after the command name.
-    pub params: Vec<String>,
-    /// Username of the invoker.
-    pub sender: String,
-    /// Room ID.
-    pub room_id: String,
-    /// Message ID that triggered this command.
-    pub message_id: String,
-    /// Timestamp of the triggering message.
-    pub timestamp: DateTime<Utc>,
-    /// Scoped handle for reading chat history.
-    pub history: HistoryReader,
-    /// Scoped handle for writing back to the chat.
-    pub writer: ChatWriter,
-    /// Snapshot of room metadata.
-    pub metadata: RoomMetadata,
-    /// All registered commands (so `/help` can list them without
-    /// holding a reference to the registry).
-    pub available_commands: Vec<CommandInfo>,
-}
-
-// ── PluginResult ────────────────────────────────────────────────────────────
-
-/// What the broker should do after a plugin handles a command.
-pub enum PluginResult {
-    /// Send a private reply only to the invoker.
-    Reply(String),
-    /// Broadcast a message to the entire room.
-    Broadcast(String),
-    /// Command handled silently (side effects already done via [`ChatWriter`]).
-    Handled,
-}
-
-// ── HistoryReader ───────────────────────────────────────────────────────────
-
-/// Scoped read-only handle to a room's chat history.
-///
-/// Respects DM visibility — a plugin invoked by user X will not see DMs
-/// between Y and Z.
-pub struct HistoryReader {
-    chat_path: PathBuf,
-    viewer: String,
-}
-
-impl HistoryReader {
-    pub(crate) fn new(chat_path: &Path, viewer: &str) -> Self {
-        Self {
-            chat_path: chat_path.to_owned(),
-            viewer: viewer.to_owned(),
-        }
-    }
-
-    /// Load all messages (filtered by DM visibility).
-    pub async fn all(&self) -> anyhow::Result<Vec<Message>> {
-        let all = history::load(&self.chat_path).await?;
-        Ok(self.filter_dms(all))
-    }
-
-    /// Load the last `n` messages (filtered by DM visibility).
-    pub async fn tail(&self, n: usize) -> anyhow::Result<Vec<Message>> {
-        let all = history::tail(&self.chat_path, n).await?;
-        Ok(self.filter_dms(all))
-    }
-
-    /// Load messages after the one with the given ID (filtered by DM visibility).
-    pub async fn since(&self, message_id: &str) -> anyhow::Result<Vec<Message>> {
-        let all = history::load(&self.chat_path).await?;
-        let start = all
-            .iter()
-            .position(|m| m.id() == message_id)
-            .map(|i| i + 1)
-            .unwrap_or(0);
-        Ok(self.filter_dms(all[start..].to_vec()))
-    }
-
-    /// Count total messages in the chat.
-    pub async fn count(&self) -> anyhow::Result<usize> {
-        let all = history::load(&self.chat_path).await?;
-        Ok(all.len())
-    }
-
-    fn filter_dms(&self, messages: Vec<Message>) -> Vec<Message> {
-        messages
-            .into_iter()
-            .filter(|m| match m {
-                Message::DirectMessage { user, to, .. } => {
-                    user == &self.viewer || to == &self.viewer
-                }
-                _ => true,
-            })
-            .collect()
-    }
-}
-
-// ── ChatWriter ──────────────────────────────────────────────────────────────
-
-/// Short-lived scoped handle for a plugin to write messages to the chat.
-///
-/// Posts as `plugin:<name>` — plugins cannot impersonate users. The writer
-/// is valid only for the duration of [`Plugin::handle`].
-pub struct ChatWriter {
-    clients: ClientMap,
-    chat_path: Arc<PathBuf>,
-    room_id: Arc<String>,
-    seq_counter: Arc<AtomicU64>,
-    /// Identity the writer posts as (e.g. `"plugin:stats"`).
-    identity: String,
-}
-
-impl ChatWriter {
-    pub(crate) fn new(
-        clients: &ClientMap,
-        chat_path: &Arc<PathBuf>,
-        room_id: &Arc<String>,
-        seq_counter: &Arc<AtomicU64>,
-        plugin_name: &str,
-    ) -> Self {
-        Self {
-            clients: clients.clone(),
-            chat_path: chat_path.clone(),
-            room_id: room_id.clone(),
-            seq_counter: seq_counter.clone(),
-            identity: format!("plugin:{plugin_name}"),
-        }
-    }
-
-    /// Broadcast a system message to all connected clients and persist to history.
-    pub async fn broadcast(&self, content: &str) -> anyhow::Result<()> {
-        let msg = make_system(&self.room_id, &self.identity, content);
-        broadcast_and_persist(&msg, &self.clients, &self.chat_path, &self.seq_counter).await?;
-        Ok(())
-    }
-
-    /// Send a private system message only to a specific user.
-    pub async fn reply_to(&self, username: &str, content: &str) -> anyhow::Result<()> {
-        let msg = make_system(&self.room_id, &self.identity, content);
-        let seq = self.seq_counter.fetch_add(1, Ordering::SeqCst) + 1;
-        let mut msg = msg;
-        msg.set_seq(seq);
-        history::append(&self.chat_path, &msg).await?;
-
-        let line = format!("{}\n", serde_json::to_string(&msg)?);
-        let map = self.clients.lock().await;
-        for (uname, tx) in map.values() {
-            if uname == username {
-                let _ = tx.send(line.clone());
-            }
-        }
-        Ok(())
-    }
-
-    /// Broadcast a typed event to all connected clients and persist to history.
-    pub async fn emit_event(
-        &self,
-        event_type: EventType,
-        content: &str,
-        params: Option<serde_json::Value>,
-    ) -> anyhow::Result<()> {
-        let msg = make_event(&self.room_id, &self.identity, event_type, content, params);
-        broadcast_and_persist(&msg, &self.clients, &self.chat_path, &self.seq_counter).await?;
-        Ok(())
-    }
-}
-
-// ── RoomMetadata ────────────────────────────────────────────────────────────
-
-/// Frozen snapshot of room state for plugin consumption.
-pub struct RoomMetadata {
-    /// Users currently online with their status.
-    pub online_users: Vec<UserInfo>,
-    /// Username of the room host.
-    pub host: Option<String>,
-    /// Total messages in the chat file.
-    pub message_count: usize,
-}
-
-/// A user's online presence.
-pub struct UserInfo {
-    pub username: String,
-    pub status: String,
-}
-
-impl RoomMetadata {
-    pub(crate) async fn snapshot(
-        status_map: &StatusMap,
-        host_user: &Arc<tokio::sync::Mutex<Option<String>>>,
-        chat_path: &Path,
-    ) -> Self {
-        let map = status_map.lock().await;
-        let online_users: Vec<UserInfo> = map
-            .iter()
-            .map(|(u, s)| UserInfo {
-                username: u.clone(),
-                status: s.clone(),
-            })
-            .collect();
-        drop(map);
-
-        let host = host_user.lock().await.clone();
-
-        let message_count = history::load(chat_path)
-            .await
-            .map(|msgs| msgs.len())
-            .unwrap_or(0);
-
-        Self {
-            online_users,
-            host,
-            message_count,
-        }
-    }
-}
+// Re-export concrete bridge types. ChatWriter and HistoryReader are public
+// (used in tests and by broker/commands.rs). snapshot_metadata is crate-only.
+pub(crate) use bridge::snapshot_metadata;
+pub use bridge::{ChatWriter, HistoryReader};
 
 // ── PluginRegistry ──────────────────────────────────────────────────────────
 
@@ -858,47 +539,7 @@ mod tests {
         }
     }
 
-    // ── ParamSchema / ParamType tests ───────────────────────────────────────
-
-    #[test]
-    fn param_type_choice_equality() {
-        let a = ParamType::Choice(vec!["x".to_owned(), "y".to_owned()]);
-        let b = ParamType::Choice(vec!["x".to_owned(), "y".to_owned()]);
-        assert_eq!(a, b);
-        let c = ParamType::Choice(vec!["x".to_owned()]);
-        assert_ne!(a, c);
-    }
-
-    #[test]
-    fn param_type_number_equality() {
-        let a = ParamType::Number {
-            min: Some(1),
-            max: Some(100),
-        };
-        let b = ParamType::Number {
-            min: Some(1),
-            max: Some(100),
-        };
-        assert_eq!(a, b);
-        let c = ParamType::Number {
-            min: None,
-            max: None,
-        };
-        assert_ne!(a, c);
-    }
-
-    #[test]
-    fn param_type_variants_are_distinct() {
-        assert_ne!(ParamType::Text, ParamType::Username);
-        assert_ne!(
-            ParamType::Text,
-            ParamType::Number {
-                min: None,
-                max: None
-            }
-        );
-        assert_ne!(ParamType::Text, ParamType::Choice(vec!["a".to_owned()]));
-    }
+    // ParamType tests moved to room-protocol — only room-cli-specific tests remain here.
 
     // ── builtin_command_infos tests ───────────────────────────────────────
 
@@ -984,67 +625,7 @@ mod tests {
         assert_eq!(before, names.len(), "duplicate command names found");
     }
 
-    #[tokio::test]
-    async fn history_reader_filters_dms() {
-        let tmp = tempfile::NamedTempFile::new().unwrap();
-        let path = tmp.path();
-
-        // Write a DM between alice and bob, and a public message
-        let dm = crate::message::make_dm("r", "alice", "bob", "secret");
-        let public = crate::message::make_message("r", "carol", "hello all");
-        history::append(path, &dm).await.unwrap();
-        history::append(path, &public).await.unwrap();
-
-        // alice sees both
-        let reader_alice = HistoryReader::new(path, "alice");
-        let msgs = reader_alice.all().await.unwrap();
-        assert_eq!(msgs.len(), 2);
-
-        // carol sees only the public message
-        let reader_carol = HistoryReader::new(path, "carol");
-        let msgs = reader_carol.all().await.unwrap();
-        assert_eq!(msgs.len(), 1);
-        assert_eq!(msgs[0].user(), "carol");
-    }
-
-    #[tokio::test]
-    async fn history_reader_tail_and_count() {
-        let tmp = tempfile::NamedTempFile::new().unwrap();
-        let path = tmp.path();
-
-        for i in 0..5 {
-            history::append(
-                path,
-                &crate::message::make_message("r", "u", format!("msg {i}")),
-            )
-            .await
-            .unwrap();
-        }
-
-        let reader = HistoryReader::new(path, "u");
-        assert_eq!(reader.count().await.unwrap(), 5);
-
-        let tail = reader.tail(3).await.unwrap();
-        assert_eq!(tail.len(), 3);
-    }
-
-    #[tokio::test]
-    async fn history_reader_since() {
-        let tmp = tempfile::NamedTempFile::new().unwrap();
-        let path = tmp.path();
-
-        let msg1 = crate::message::make_message("r", "u", "first");
-        let msg2 = crate::message::make_message("r", "u", "second");
-        let msg3 = crate::message::make_message("r", "u", "third");
-        let id1 = msg1.id().to_owned();
-        history::append(path, &msg1).await.unwrap();
-        history::append(path, &msg2).await.unwrap();
-        history::append(path, &msg3).await.unwrap();
-
-        let reader = HistoryReader::new(path, "u");
-        let since = reader.since(&id1).await.unwrap();
-        assert_eq!(since.len(), 2);
-    }
+    // HistoryReader tests live in bridge.rs alongside the implementation.
 
     // ── Plugin trait default methods ──────────────────────────────────────
