+++ crates/room-protocol/Cargo.toml
@@ -7,6 +7,7 @@ license = "MIT"
 repository = "https://github.com/knoxio/room"
 
 [dependencies]
+anyhow = "1"
 chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
